renames HomepageSearchAutocomplete files to SearchAutocomplete
