+++ src/screens/HomePage/HomePage.tsx
@@ -26,7 +26,7 @@ import {
   Section,
   ColumnCentered,
 } from './HomePage.style';
-import { HomepageSearchAutocomplete } from 'components/Search';
+import SearchAutocomplete from 'components/Search';
 import Toggle from './Toggle/Toggle';
 import HorizontalThermometer from 'components/HorizontalThermometer';
 import HomepageItems from 'components/RegionItem/HomepageItems';
@@ -128,7 +128,7 @@ export default function HomePage() {
         <div className="App">
           <Content>
             <ColumnCentered id="search">
-              <HomepageSearchAutocomplete
+              <SearchAutocomplete
                 locations={searchLocations}
                 filterLimit={getFilterLimit()}
                 menuOpen={menuOpen}
